requirements.txt
asn1crypto==0.24.0
certifi==2018.10.15
cffi==1.11.5
chardet==3.0.4
cryptography==2.3.1
future==0.17.1
geojson==2.3.0
idna==2.7
oauthlib==2.0.1
pycparser==2.19
pyowm==2.9.0
python-telegram-bot==11.1.0
requests==2.20.1
requests-oauthlib==0.7.0
six==1.10.0
unsplash==1.0.0a2
urllib3==1.24.1
